pyproject.toml
[project]
name = "deadbolt"
version = "1.0.0"
description = "Lean professional pentest orchestrator"
requires-python = ">=3.10"

dependencies = [
  "typer>=0.9",
  "rich>=13",
  "pydantic>=2",
  "PyYAML>=6",
  "Jinja2>=3",
  "requests>=2.31"
]

[project.scripts]
deadbolt = "main.cli.app:app"

[build-system]
requires = ["setuptools>=65", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
packages = ["main"]